pyproject.toml
[project]
name = "moyu-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fire>=0.7.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-core>=0.3.75",
    "langchain-deepseek>=0.1.4",
    "python-dotenv>=1.1.1",
]
